mod/#21 wish -> 금화 미션 뷰로 이동
