+++ app/src/main/java/com/kiero/presentation/kid/mission/KidMissionViewModel.kt
@@ -1,8 +1,8 @@
-package com.kiero.presentation.kid.wish
+package com.kiero.presentation.kid.mission
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kiero.presentation.kid.wish.state.KidWishState
+import com.kiero.presentation.kid.mission.state.KidMissionState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.collections.immutable.toPersistentList
@@ -13,9 +13,9 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class KidWishViewModel @Inject constructor(
+class KidMissionViewModel @Inject constructor(
 ) : ViewModel() {
-    private val _state = MutableStateFlow(KidWishState())
+    private val _state = MutableStateFlow(KidMissionState())
     val state = _state.asStateFlow()
