+++ app/src/main/java/com/kiero/presentation/kid/mission/model/KidMissionUiModel.kt
@@ -1,9 +1,9 @@
-package com.kiero.presentation.kid.wish.model
+package com.kiero.presentation.kid.mission.model
 
 import androidx.compose.runtime.Immutable
 
 @Immutable
-data class KidWishMission(
+data class KidMissionUiModel(
     val id: Int = -1,
     val name: String = "", // 미션 이름
     val reward: Int = -1,
